automation/ppt_automation/requirements.txt
python-pptx>=0.6.21
openpyxl>=3.0.0
pandas>=2.0.0
xlrd>=2.0.1
